Added logging dependencies, split files
