+++ src/directory_flattener/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+log = "0.4.14"
+env_logger = "0.9.0"
 serde_json = "1.0.79"
 uuid = {version = "0.8.2", features = ["v4"]}
 walkdir = "2.3.2"
